requirements.txt
agora_token_builder==1.0.0
asarPy==1.0.1
asgiref==3.8.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
Django==4.1
django-crispy-forms==1.14.0
django-embed-video==1.4.5
django-storages==1.13.1
dlib==19.24.8
face-recognition==1.3.0
face_recognition_models==0.3.0
folium==0.19.5
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
pandas==2.2.3
Pillow==10.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
ratelim==0.1.6
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
xyzservices==2025.1.0
gunicorn==23.0.0
whitenoise==6.9.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
python-dotenv==1.0.1